requirements.txt
pytest==8.3.5
httpx==0.27.2
respx==0.21.1
python-dotenv==1.0.1
